requirements.txt
# Base requirements
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.18.5
opencv-python>=4.1.2
matplotlib>=3.2.2
pillow>=7.1.2
scipy>=1.4.1
tqdm>=4.41.0

# Enhanced metrics and visualization
seaborn>=0.11.0
pandas>=1.1.4
pyyaml>=5.3.1
tensorboard>=2.4.1
thop>=0.1.1  # FLOPs computation
albumentations>=1.0.3  # Advanced augmentations

# Export dependencies
onnx>=1.9.0  # For ONNX export
onnxruntime>=1.8.0
onnx-simplifier>=0.3.6

# Optional for GIF support
imageio>=2.9.0
imageio-ffmpeg>=0.4.5